Move cfg and check-cfg configuration in it's own module and add docs
